feat: [ga2] select / create a logo #756)
