+++ app/theme/common/palette.ts
@@ -9,14 +9,6 @@ const CAUTION = {
 };
 
 /**
- * Palette "Primary"
- */
-const PRIMARY = {
-  DARK: "#1F1F47",
-  MAIN: "#28285B",
-};
-
-/**
  * Palette Option "Caution"
  */
 const caution: PaletteColorOptions = {
@@ -25,18 +17,8 @@ const caution: PaletteColorOptions = {
 };
 
 /**
- * Palette Option "Primary"
- */
-const primary: PaletteColorOptions = {
-  contrastText: "#FFFFFF",
-  dark: PRIMARY.DARK,
-  main: PRIMARY.MAIN,
-};
-
-/**
  * App Palette
  */
 export const palette: PaletteOptions = {
   caution,
-  primary,
 };
